feat(lib): maz-carousel - new component
