+++ packages/lib/package/components/index.ts
@@ -7,6 +7,7 @@ import { default as MazBadge } from './MazBadge.vue'
 import { default as MazBottomSheet } from './MazBottomSheet.vue'
 import { default as MazBtn } from './MazBtn.vue'
 import { default as MazCard } from './MazCard.vue'
+import { default as MazCarousel } from './MazCarousel.vue'
 import { default as MazChart } from './MazChart.vue'
 import { default as MazCheckbox } from './MazCheckbox.vue'
 import { default as MazDialog } from './MazDialog.vue'
@@ -39,6 +40,7 @@ export {
   MazBottomSheet,
   MazBtn,
   MazCard,
+  MazCarousel,
   MazChart,
   MazCheckbox,
   MazDialog,
@@ -72,6 +74,7 @@ export default {
   MazBottomSheet,
   MazBtn,
   MazCard,
+  MazCarousel,
   MazChart,
   MazCheckbox,
   MazDialog,
